Moved everything from gameloop into quadtree
now we just run quadtrees initialize and update in gameloop.
Linus added smartpointers but they aren't used yet.
Drawable now has a bounding box.
Drawables bounding box is always on the same position with the same size.
We need to add the ability to change the pos and size depending on where and how big the meshInstance is.