Worked on render logic
drawables bounding box belongs to meshInstance now.
Matthew wants to add functionality to recreate the quadtree IF something in it has moved. (this is not a needed functionality for the game to work).
We want some way to seperate all the interesting dynamic objects so we only render those dynamic objects and the visible static objects for any frame.